index-photos/requirements.txt
boto3
requests
python-magic